models/explorer.js created/ models/planets.js updated
